deploy/requirements.txt
absl-py==0.7.1
astor==0.7.1
audioread==2.1.6
backcall==0.1.0
bleach==2.1.4
boto3==1.9.151
botocore==1.12.151
cachetools==2.1.0
certifi==2018.8.24
cffi==1.13.1
chardet==3.0.4
click==6.7
gunicorn==19.6.0
colorama==0.3.9
cx-Freeze==5.1.1
cycler==0.10.0
Cython==0.28.5
decorator==4.3.0
distro==1.4.0
dnspython==1.16.0
docutils==0.14
entrypoints==0.2.3
Flask==1.0.2
Flask-Compress==1.4.0
future==0.16.0
gast==0.2.2
gevent==1.4.0
google-api-core==1.14.3
google-auth==1.6.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.0.3
google-pasta==0.1.7
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
graphviz==0.10.1
greenlet==0.4.15
grpcio==1.20.0
gym==0.10.5
h5py==2.9.0
html5lib==1.0.1
idna==2.7
image==1.5.27
imageio==2.5.0
ipyevents==0.3.1
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
jmespath==0.9.3
joblib==0.13.0
jsonschema==2.6.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
llvmlite==0.27.0
Markdown==3.1
MarkupSafe==1.0
mistune==0.8.3
mock==2.0.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.3
notebook==5.6.0
numba==0.42.0
numpy==1.16.1
oauthlib==3.1.0
opencv-python==4.1.0.25
pandocfilters==1.4.2
parso==0.3.1
patsy==0.5.0
pbr==5.1.3
pickleshare==0.7.4
Pillow==6.0.0
plotly==3.2.1
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.7.1
psycopg2-binary==2.7.5
py2exe==0.9.2.2
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pydata-google-auth==0.1.3
pydot==1.4.1
pydotplus==2.0.2
pyglet==1.3.2
Pygments==2.2.0
pymongo==3.7.2
pyparsing==2.2.0
pystan==2.18.0.0
pytesseract==0.3.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.3
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
requests==2.19.1
requests-oauthlib==1.2.0
resampy==0.2.1
retrying==1.3.3
rsa==3.4.2
s3transfer==0.2.0
schedule==0.5.0
scikit-image==0.15.0
scikit-learn==0.19.2
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.12
sqlparse==0.3.0
statsmodels==0.9.0
superintendent==0.2.3
tb-nightly==1.14.0a20190301
tensorflow==2.0.0a0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tf-estimator-nightly==1.14.0.dev2019030115
tornado==5.1
traitlets==4.3.2
urllib3==1.23
virtualenv==16.7.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.1
wrapt==1.11.2
xlrd==1.2.0
